requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
APScheduler==3.11.2
asgiref==3.11.0
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
cloudinary==1.44.1
dj-database-url==3.1.0
Django==5.2.9
django-apscheduler==0.7.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-sslserver==0.22
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
frozenlist==1.8.0
geographiclib==2.1
geopy==2.4.1
gunicorn==23.0.0
idna==3.11
multidict==6.7.1
packaging==25.0
pillow==12.1.0
propcache==0.4.1
PyJWT==2.10.1
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
twilio==9.10.0
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
whitenoise==6.11.0
yarl==1.22.0
